+++ internal/xray/tun_other.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package xray
+
+func prepareTUN() error {
+	return nil
+}